v_t

+++ mmaction/models/matchers/video_text_matcher_memory_bank_e2e.py
@@ -140,7 +140,9 @@ class VideoTextMatcherBankE2E(BaseMatcher):
             v_feat, t_feat = self.neck(v_feat, t_feat)
 
         slct_idx = (
-            torch.zeros(batch_size, self.bank_size + 1).to(device).detach()
+            torch.zeros(batch_size, self.bank_size + 1, dtype=torch.long)
+            .to(device)
+            .detach()
         )  # [batch_size, bank_size + 1]
         for i, idx in enumerate(idxes):
             self.probs[idx] = 0.0
